Remove 'Resource Group' from Networking list
A resource group is not a networking component of VPC, it's a construct of IAM that separates resources and allows permissioned access. Different resources in a VPC can exist in different Resource Groups too. Resource Groups are not referred to again in the text below the table, so I think it should be removed to avoid confusion.